cast bool to int track label

+++ src/graphnet/training/labels.py
@@ -103,4 +103,4 @@ class Track(Label):
     def __call__(self, graph: Data) -> torch.tensor:
         """Compute label for `graph`."""
         label = (graph[self._pid_key] == 14) & (graph[self._int_key] == 1)
-        return label  # torch.tensor(label, dtype=torch.int64)
+        return label.type(torch.int)  # torch.tensor(label, dtype=torch.int64)
